+++ tests/ClassNotation/_data/AbstractA.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace ErickSkrauch\PhpCsFixer\Tests\ClassNotation\_data;
+
+abstract class AbstractA implements InterfaceA {
+
+    public function bar(): void {
+    }
+
+}