+++ tests/ClassNotation/_data/EmptyInterface.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace ErickSkrauch\PhpCsFixer\Tests\ClassNotation\_data;
+
+interface EmptyInterface {
+
+}