+++ tests/ClassNotation/_data/InterfaceA.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace ErickSkrauch\PhpCsFixer\Tests\ClassNotation\_data;
+
+interface InterfaceA {
+
+    public function foo(): void;
+
+    public function bar(): void;
+
+}