+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Enh #4: Introduce `ErickSkrauch\ordered_overrides` fixer.
 
 ## [1.2.4] - 2024-01-15
 ### Fixed